+++ src/Aardvark.Rendering.Vulkan/Runtime/Headless.fs
@@ -3,6 +3,7 @@
 open System
 open Aardvark.Base
 open Aardvark.Rendering
+open System.Runtime.InteropServices
 
 type HeadlessVulkanApplication(debug : DebugLevel, instanceExtensions : list<string>, deviceExtensions : PhysicalDevice -> list<string>) =
     let requestedExtensions =
@@ -15,17 +16,31 @@ type HeadlessVulkanApplication(debug : DebugLevel, instanceExtensions : list<str
 
             yield! Instance.Extensions.Raytracing
 
-            yield KHRVideoQueue.Name
-            yield KHRVideoDecodeQueue.Name
-            yield KHRVideoEncodeQueue.Name
-            yield EXTVideoDecodeH264.Name
-            yield EXTVideoEncodeH264.Name
-            yield EXTVideoDecodeH265.Name
-            yield EXTVideoEncodeH265.Name
 
             if debug > DebugLevel.None then
                 yield Instance.Extensions.DebugReport
                 yield Instance.Extensions.DebugUtils
+
+            yield KHRGetPhysicalDeviceProperties2.Name
+            yield KHRExternalMemoryCapabilities.Name
+            yield KHRExternalMemory.Name
+            yield KHRExternalFenceCapabilities.Name
+            yield KHRExternalFence.Name
+            yield KHRExternalSemaphoreCapabilities.Name
+            yield KHRExternalSemaphore.Name
+            yield EXTExternalMemoryHost.Name
+            yield EXTExternalMemoryDmaBuf.Name
+
+            if RuntimeInformation.IsOSPlatform OSPlatform.Windows then
+                yield KHRExternalMemoryWin32.Name
+                yield KHRExternalFenceWin32.Name
+                yield KHRExternalSemaphoreWin32.Name
+
+            elif RuntimeInformation.IsOSPlatform OSPlatform.Linux then
+                yield KHRExternalMemoryFd.Name
+                yield KHRExternalFenceFd.Name
+                yield KHRExternalSemaphoreFd.Name
+
         ]
 
     let requestedLayers =